# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API sample",
    "version": "1.0.0",
    "description": "This is a sample API.",
    "contact": {
      "email": "admin@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://example.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "develop"
    },
    {
      "url": "https://example.com/api",
      "description": "production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
    "/auth/access_token": {
      "post": {
        "summary": "issue access_token",
        "description": "issue access_token by userId and api secret",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "unique user id",
                    "example": "clqj40goe0000z21ibpk26izw"
                  },
                  "secret": {
                    "type": "string",
                    "description": "api secret",
                    "example": "YOUR_SECRET"
                  }
                },
                "required": [
                  "userId",
                  "secret"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {
                      "type": "string",
                      "description": "access_token",
                      "example": "eyJhbGciOiJIUzI1NiJ9.eyJ1c2VySWQiOjEsInVzZXJOYW1lIjoiYWRtaW4iLCJpbWFnZVVybCI6Ii90YWt0LnBuZyIsImlhdCI6MTcwMzA0MDUzMCwiaXNzIjoidXJuOmV4YW1wbGU6aXNzdWVyIiwiYXVkIjoidXJuOmV4YW1wbGU6YXVkaWVuY2UiLCJleHAiOjE3MDMwNzY1MzB9.UGcaC4PI-_2hFTThL3WhgTrMZbnulivAoyVjh14hq-d"
                    }
                  },
                  "required": [
                    "access_token"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Bad Request"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Unauthorized"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Internal Server Error"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "path": {
                                "type": "string",
                                "example": "error path"
                              },
                              "message": {
                                "type": "string",
                                "example": "error message"
                              }
                            },
                            "required": [
                              "path",
                              "message"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}